% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{add_inventory}
\alias{add_inventory}
\title{Upload a CSV of listings to add to your inventory.}
\usage{
add_inventory(
 csv,
 token
)
}
\arguments{
\item{csv}{A string representing the source location of a csv file, see Details}

\item{token}{Token object obtained from authorize()}
}
\description{
Upload a CSV of listings to add to your inventory.
}
\details{
\preformatted{   release_id,price,media_condition,comments
   123,1.50,Mint (M),Comments about this release for sale
   456,2.50,Near Mint (NM or M-),More comments
   7890,3.50,Good (G),Signed vinyl copy

``
}
}
\examples{
token <- authorize("key", "secret")
}
