% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{add_order_message}
\alias{add_order_message}
\title{Add a new message to the order's message log}
\usage{
add_order_message(
  order_id
  token = NA,
  options = list(),
)
}
\arguments{
\item{order_id}{The ID of the order you are fetching}

\item{token}{(optional) Token object obtained from authorize()}

\item{options}{(optional) Parameters, see Details}
}
\description{
Add a new message to the order's message log
}
\details{
Adds a new message to the order’s message log.
When posting a new message, you can simultaneously change the order status. If you do, the message will automatically be prepended with:
Seller changed status from Old Status to New Status
While message and status are each optional, one or both must be present.
}
