% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{add_to_collection_folder}
\alias{add_to_collection_folder}
\title{Add release to collection}
\usage{
add_to_collection_folder(
 username,
 folder_id,
 release_id,
 token
)
}
\arguments{
\item{username}{String containing a valid username}

\item{folder_id}{Valid identifier for a folder}

\item{release_id}{Integer value representing a valid release ID}

\item{token}{Token object obtained from authorize()}
}
\description{
Add release to collection
}
\details{
Add a release to a folder in a user’s collection.
The folder_id must be non-zero – you can use 1 for “Uncategorized”.
Authentication as the collection owner is required.
}
\examples{
token <- authorize("key", "secret")
}
