% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wantlist.R
\name{add_to_wantlist}
\alias{add_to_wantlist}
\title{Add a release to your wantlist}
\usage{
add_to_wantlist(
 username,
 release_id,
 token,
 options = list()
)
}
\arguments{
\item{username}{String containing a valid username}

\item{release_id}{Integer value representing a valid release ID}

\item{token}{Token object obtained from authorize()}

\item{options}{(optional) List of named parameters, see Details}
}
\description{
Add a release to your wantlist
}
\details{
Add a release to a user’s wantlist.
Authentication as the wantlist owner is required.
}
\examples{
token <- authorize("key", "secret")
}
