% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{create_collection_folder}
\alias{create_collection_folder}
\title{Create a new folder in your collection}
\usage{
create_collection_folder(
 username,
 token,
 name = ""
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}

\item{name}{}
}
\description{
Create a new folder in your collection
}
\details{
Authentication as the collection owner is required.
}
\examples{
#'token <- authorize("key", "secret")
}
