% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{delete}
\alias{delete}
\title{Delete request to the discogs server}
\usage{
delete(
 url,
 token
)
}
\arguments{
\item{url}{full URL to API endpoint}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Delete request to the discogs server
}
\details{
This function is not supposed to used directly, but through other functions instead
}
