% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{delete_collection_folder}
\alias{delete_collection_folder}
\title{Delete a folder from a user’s collection.}
\usage{
delete_collection_folder(
 username,
 folder_id
 token,
)
}
\arguments{
\item{username}{String containing a valid username}

\item{folder_id}{Valid identifier for a folder}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Delete a folder from a user’s collection.
}
\details{
A folder must be empty before it can be deleted.
Authentication as the collection owner is required.
}
\examples{
token <- authorize("key", "secret")
delete_collection_folder("username", 1, token)
}
