% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wantlist.R
\name{delete_from_wantlist}
\alias{delete_from_wantlist}
\title{Delete a release from your wantlist}
\usage{
delete_from_wantlist(
 username,
 release_id,
 token,
 options = list()
)
}
\arguments{
\item{username}{String containing a valid username}

\item{release_id}{Integer value representing a valid release ID}

\item{token}{Token object obtained from authorize()}

\item{options}{(optional) List of named parameters, see Details}
}
\description{
Delete a release from your wantlist
}
\details{

}
\examples{
token <- authorize("key", "secret")
}
