% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{delete_inventory}
\alias{delete_inventory}
\title{Upload a CSV of listings to delete from your inventory.}
\usage{
delete_inventroy(
 csv,
 token
)
}
\arguments{
\item{csv}{A string representing the source location of a csv file, see Details}

\item{token}{Token object obtained from authorize()}
}
\description{
Upload a CSV of listings to delete from your inventory.
}
\details{

}
\examples{
token <- authorize("key", "secret")
}
