% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{delete_listing}
\alias{delete_listing}
\title{Delete a listing from the Marketplace}
\usage{
delete_listing(
  listing_id,
  token,
  options = list(),
)
}
\arguments{
\item{listing_id}{The ID of the listing you are fetching}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{options}{(optional) Parameters, see Details}
}
\description{
Delete a listing from the Marketplace
}
\details{
Permanently remove a listing from the Marketplace.
Authentication as the listing owner is required.
}
