% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{delete_release_from_folder}
\alias{delete_release_from_folder}
\title{Delete Instance From Folder}
\usage{
delete_release_from_folder(
 username,
 folder_id,
 release_id,
 instance_id,
 token
)
}
\arguments{
\item{username}{String containing a valid username}

\item{folder_id}{Valid identifier for a folder}

\item{release_id}{Integer value representing a valid release ID}

\item{instance_id}{The ID of the instance}

\item{token}{Token object obtained from authorize()}
}
\description{
Delete Instance From Folder
}
\details{
Remove an instance of a release from a user’s collection folder.

To move the release to the “Uncategorized” folder instead, use add_to_collection()

Authentication as the collection owner is required.
}
\examples{
token <- authorize("key", "secret")
}
