% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{delete_release_rating}
\alias{delete_release_rating}
\title{Delete a given rating}
\usage{
delete_release_rating(
 release_id,
 username,
 token
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}
}
\description{
Delete a given rating
}
\examples{
# Delete a rating
token <- authorize("key", "secret")
delete_release_rating(1, "username", 5, token)
}
