% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{edit_collection_folder}
\alias{edit_collection_folder}
\title{Edit a folder’s metadata.}
\usage{
edit_collection_folder(
 username,
 folder_id
 token,
 name = ""
)
}
\arguments{
\item{username}{String containing a valid username}

\item{folder_id}{Valid identifier for a folder}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{name}{(optional) Name for the folder}
}
\description{
Edit a folder’s metadata.
}
\details{
Folders 0 and 1 cannot be renamed.
Authentication as the collection owner is required.
}
\examples{
token <- authorize("key", "secret")
edit_collection_folder("username", 3, token, name = "new_folder_name")
}
