% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{edit_listing}
\alias{edit_listing}
\title{Edit the data associated with a listing.}
\usage{
post_listing(
  listing_id,
  release_id,
  condition,
  price,
  status,
  token,
  options = list()
)
}
\arguments{
\item{listing_id}{The ID of the listing you are fetching}

\item{release_id}{Integer value representing a valid release ID}

\item{condition}{The condition of the release you are posting. See Details}

\item{price}{The price of the item (in the seller’s currency).}

\item{status}{The status of the listing. See Details}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{options}{(optional) Parameters, see Details}
}
\description{
Edit the data associated with a listing.
}
\details{
Create a Marketplace listing.
Authentication is required; the listing will be added to the authenticated user’s Inventory.
\subsection{Condition}{

Condition must be one of the following:
\itemize{
\item \strong{Mint(M)}
\item \strong{Near Mint (NM or M-)}
\item \strong{Very Good Plus (VG+)}
\item \strong{Very Good (VG)}
\item \strong{Good Plus (G+)}
\item \strong{Good (G)}
\item \strong{Fair (F)}
\item \strong{Poor (P)}
}
}

\subsection{Status}{

Defaults to \strong{For Sale}. Options are:
\itemize{
\item \strong{For Sale} (the listing is ready to be shown on the Marketplace)
\item \strong{Draft} (the listing is not ready for public display).
}
}

\subsection{Parameters}{

Below are optional parameters that can be provided to the options argument
\subsection{sleeve_condition}{

The condition of the sleeve of the item you are posting. Must be one of the following:
\itemize{
\item \strong{Mint(M)}
\item \strong{Near Mint (NM or M-)}
\item \strong{Very Good Plus (VG+)}
\item \strong{Very Good (VG)}
\item \strong{Good Plus (G+)}
\item \strong{Good (G)}
\item \strong{Fair (F)}
\item \strong{Poor (P)}
\item \strong{Generic}
\item \strong{Not Graded}
\item \strong{No Cover}
}
}

\subsection{comments}{

Any remarks about the item that will be displayed to buyers.
}

\subsection{allow_offers}{

Whether or not to allow buyers to make offers on the item. Defaults to \strong{False}.
}

\subsection{external_id}{

A freeform field that can be used for the seller’s own reference.
Information stored here will not be displayed to anyone other than the seller.
This field is called “Private Comments” on the Discogs website.
}

\subsection{location}{

A freeform field that is intended to help identify an item’s physical storage location.
Information stored here will not be displayed to anyone other than the seller.
This field will be visible on the inventory management page and will be available in inventory exports via the website.
}

\subsection{weight}{

The weight, in grams, of this listing, for the purpose of calculating shipping.
Set this field to \strong{auto} to have the weight automatically estimated for you.
}

\subsection{format_quantity}{

The number of items this listing counts as, for the purpose of calculating shipping.
This field is called “Counts As” on the Discogs website.
Set this field to \strong{auto} to have the quantity automatically estimated for you.
}

}
}
\examples{

token <- authorize("key", "secret")
edit_listing(1, 10, "Near Mint (NM or M-)", "10.30", "For Sale", token)

edit_listing(13, 112301, "Near Mint (NM or M-)", 12.30, "Draft", token, options=list("weight" = 400))
}
