% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{edit_order}
\alias{edit_order}
\title{Edit the data associated with an order}
\usage{
edit_order(
  order_id,
  token,
  options = list()
)
}
\arguments{
\item{order_id}{The ID of the order you are fetching}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{options}{(optional) Parameters, see Details}
}
\description{
Edit the data associated with an order
}
\details{
Edit the data associated with an order.
Authentication as the seller is required.
The response contains a next_status key – an array of valid next statuses for this order, which you can display to the user in (for example) a dropdown control.
This also renders your application more resilient to any future changes in the order status logic.
Changing the order status using this resource will always message the buyer with:\preformatted{_Seller changed status from Old Status to New Status_
}

and does not provide a facility for including a custom message along with the change.
For more fine-grained control, use the Add a new message resource, which allows you to simultaneously add a message and change the order status.
If the order status is neither cancelled, Payment Received, nor Shipped, you can change the shipping.
Doing so will send an invoice to the buyer and set the order status to Invoice Sent.
(For that reason, you cannot set the shipping and the order status in the same request.)
}
\examples{
token <- authorize("key", "secret")
edit_order(1, token)
}
