% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identity.R
\name{edit_profile}
\alias{edit_profile}
\title{Edit a user’s profile data.}
\usage{
edit_profile(
 username,
 token,
 options = list()
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}

\item{options}{(optional) Parameters, see Details}
}
\description{
Edit a user’s profile data.
}
\details{
Authentication as the user is required.
For options, the following parameters are allowed:
\itemize{
\item \strong{name}: The real name of the user.
\item \strong{home_page}: The user’s website.
\item \strong{location}: The geographical location of the user.
\item \strong{profile}: Biographical information about the user.
\item \strong{curr_abbr}: Currency for marketplace data.
}
}
\examples{
token <- authorize("key", "secret")
edit_profile("name", token, options = list("home_page" = "www.discogs.com"))
}
