% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{export_inventory}
\alias{export_inventory}
\title{Export your inventory}
\usage{
export_inventory(
 token
)
}
\arguments{
\item{token}{Token object obtained from authorize()}
}
\description{
Export your inventory
}
\details{

}
\examples{
token <- authorize("key", "secret")
export_inventory(token)
}
