% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{get}
\alias{get}
\title{get a response from a request with token}
\usage{
get(
 url,
 token = NA
)
}
\arguments{
\item{token}{Token object obtained from authorize()}
}
\description{
get a response from a request with token
}
\details{
This function is not supposed to used directly, but is used through other functions
}
