% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_artist}
\alias{get_artist}
\title{Retrieve artist information}
\usage{
get_artist(
 release_id,
 token = NA
)
}
\arguments{
\item{artist_id}{Integer value representing a valid artist ID}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Retrieve artist information
}
\examples{
# Retrieve artist information
artist <- get_artist(1)
}
