% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_artist_releases}
\alias{get_artist_releases}
\title{Get releases by artist}
\usage{
get_artist_releases(
 artist_id,
 options = list(),
 token = NA
)
}
\arguments{
\item{artist_id}{Integer value representing a valid artist ID}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get releases by artist
}
\details{
This function supports Pagination parameters. Below is a list of available fields.
\itemize{
\item{ \strong{sort:} }{string, sort items by these fields: year, title, format}
\item{ \strong{sort_order:} }{string, sort items in a particular order; \strong{example}: "asc"}
}
}
\examples{
# Retrieve the release statistics
releases <- get_artist_releases(1)

releases <- get_artist_releases(1, options=list("sort" = "year"))
}
