% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{get_collection}
\alias{get_collection}
\title{Retrieve a list of folders in a user’s collection.}
\usage{
get_collection(
 username,
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}
}
\description{
Retrieve a list of folders in a user’s collection.
}
\details{
If the collection has been made private by its owner, authentication as the collection owner is required.
If you are not authenticated as the collection owner, only folder ID 0 (the “All” folder) will be visible
(if the requested user’s collection is public).
}
\examples{
token <- authorize("key", "secret")

}
