% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{get_collection_folder}
\alias{get_collection_folder}
\title{Retrieve metadata about a folder in a user’s collection.}
\usage{
get_collection_folder(
 username,
 folder_id,
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{folder_id}{Valid identifier for a folder}

\item{token}{Token object obtained from authorize()}
}
\description{
Retrieve metadata about a folder in a user’s collection.
}
\details{
If folder_id is not 0, authentication as the collection owner is required.
}
\examples{
#'token <- authorize("key", "secret")
}
