% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{get_collection_items_by_folder}
\alias{get_collection_items_by_folder}
\title{Collection Items By Folder}
\usage{
get_collection_items_by_folder(
 username,
 folder_id,
 options = list(),
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{folder_id}{Valid identifier for a folder}

\item{options}{(optional) List of named parameters, see Details}

\item{token}{Token object obtained from authorize()}
}
\description{
Collection Items By Folder
}
\details{

}
\examples{
Returns the list of item in a folder in a user’s collection. Accepts Pagination parameters.
Basic information about each release is provided, suitable for display in a list. For detailed information, make another API call to fetch the corresponding release.
If folder_id is not 0, or the collection has been made private by its owner, authentication as the collection owner is required.
If you are not authenticated as the collection owner, only public notes fields will be visible.
Valid sort keys are: `label`, `artist`, `title`, `catno`, `format`, `rating`, `added`, `year`
}
