% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{get_collection_items_by_release}
\alias{get_collection_items_by_release}
\title{Get collection items by release}
\usage{
get_collection_items_by_release(
 username,
 release_id
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{release_id}{Integer value representing a valid release ID}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get collection items by release
}
\details{
View the user’s collection folders which contain a specified release.
This will also show information about each release instance.

The release_id must be non-zero.

Authentication as the collection owner is required if the owner’s collection is private.
}
\examples{
items <- get_collection_items_by_release("username", 1000)
}
