% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{get_collection_value}
\alias{get_collection_value}
\title{Get the value of the user's collection}
\usage{
get_collection_value(
 username,
 token
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}
}
\description{
Get the value of the user's collection
}
\details{
Returns the minimum, median, and maximum value of a user’s collection.

Authentication as the collection owner is required.
}
\examples{
token <- authorize("key", "secret")
}
