% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_community_release_rating}
\alias{get_community_release_rating}
\title{Get the rating statistics by the community}
\usage{
get_community_release_rating(
 release_id,
 token = NA
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Get the rating statistics by the community
}
\examples{
# Retrieve the community rating
rating <- get_community_release_rating(1)
}
