% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{get_export_details}
\alias{get_export_details}
\title{Get details about a specific export}
\usage{
get_export_details(
 export_id,
 token
)
}
\arguments{
\item{export_id}{Id of the export}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get details about a specific export
}
\examples{
token <- authorize("key", "secret")
details <- get_export_details(1, token)
}
