% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_fee}
\alias{get_fee}
\title{Get the fee for selling an item on the Marketplace}
\usage{
get_fee(
  options = list(),
  token = NA
)
}
\arguments{
\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Get the fee for selling an item on the Marketplace
}
\details{
The Fee resource allows you to quickly calculate the fee for selling an item on the Marketplace.
}
