% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_fee_with_currency}
\alias{get_fee_with_currency}
\title{Get the fee for your currency}
\usage{
get_fee_with_currency(
  price,
  options = list(),
  token = NA
)
}
\arguments{
\item{price}{The price to calculate a fee from}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get the fee for your currency
}
\details{
The Fee resource allows you to quickly calculate the fee for selling an item on the Marketplace given a particular currency.
}
