% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_inventory}
\alias{get_inventory}
\title{Get a seller’s inventory}
\usage{
get_inventory(
  username,
  options = list(),
  token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get a seller’s inventory
}
\details{
If you are not authenticated as the inventory owner, only items that have a status of For Sale will be visible.
If you are authenticated as the inventory owner you will get additional weight, format_quantity, external_id, and location keys.
If the user is authorized, the listing will contain a in_cart boolean field indicating whether or not this listing is in their cart.
\subsection{Parameters}{

Below are optional parameters that can be provided to the options argument
\subsection{status}{

Only show items with this status, for example: \verb{for sale}
}

\subsection{sort}{

Sort items by one of the following fields: \code{listed}, \code{price}, \code{item}  (i.e. the title of the release)
\code{artist}, \code{label}, \code{catno}, \code{audio}, \code{status} (when authenticated as the inventory owner),
\code{location} (when authenticated as the inventory owner)
}

\subsection{sort_order}{

Sort items in a particular order (one of \code{asc} or \code{desc})
}

}
}
\examples{
inventory <- get_inventory("username")

inventory <- get_inventory("username", options = list("sort" = "artist", "sort_order" = "desc"))
}
