% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_label}
\alias{get_label}
\title{Get label information}
\usage{
get_label(
 label_id,
 token = NA
)
}
\arguments{
\item{label_id}{Integer value representing a valid release ID}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get label information
}
\examples{
# Retrieve label information
label <- get_label(1)
}
