% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Userlist.R
\name{get_list}
\alias{get_list}
\title{Returns items from a specified List.}
\usage{
get_list(
 list_id,
 token = NA
)
}
\arguments{
\item{list_id}{Integer identifier for a user list}

\item{token}{Token object obtained from authorize()}
}
\description{
Returns items from a specified List.
}
\details{
Private Lists will only display when authenticated as the owner.
}
