% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{get_list_custom_fields}
\alias{get_list_custom_fields}
\title{List custom fields in collection}
\usage{
get_list_custom_fields(
 username,
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
List custom fields in collection
}
\details{
Retrieve a list of user-defined collection notes fields. These fields are available on every release in the collection.

If the collection has been made private by its owner, authentication as the collection owner is required.

If you are not authenticated as the collection owner, only fields with public set to true will be visible.
}
\examples{
token <- authorize("key", "secret")
list <- get_list_custom_fields("username", token)
}
