% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_list_order_messages}
\alias{get_list_order_messages}
\title{Get messages of a single order}
\usage{
get_list_order_messages(
  order_id,
  token,
  options = list(),
)
}
\arguments{
\item{order_id}{The ID of the order you are fetching}

\item{token}{Token object obtained from authorize()}

\item{options}{(optional) Parameters, see Details}
}
\description{
Get messages of a single order
}
\details{
Returns a list of the order’s messages with the most recent first. Accepts Pagination parameters.
Authentication as the seller is required.
}
\examples{
token <- authorize("key", "secret")
}
