% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_listing}
\alias{get_listing}
\title{View the data associated with a listing}
\usage{
get_listing(
  listing_id,
  options = list(),
  token = NA
)
}
\arguments{
\item{listing_id}{The ID of the listing you are fetching}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
View the data associated with a listing
}
\details{
If the authorized user is the listing owner the listing will include the weight, format_quantity, external_id, and location keys.
If the user is authorized, the listing will contain a in_cart boolean field indicating whether or not this listing is in their cart.
}
\examples{

listing <- get_listing(1)

listing <- get_listing(1, options = list("curr_abbr" = "EUR"))
}
