% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_master_release}
\alias{get_master_release}
\title{Get a master release}
\usage{
get_master_release(
 master_id,
 token = NA
)
}
\arguments{
\item{master_id}{Integer value representing a valid master ID}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Get a master release
}
\examples{
# Retrieve a master release
release <- get_master_release(1000)
}
