% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_master_release_versions}
\alias{get_master_release_versions}
\title{Get versions of the given master release}
\usage{
get_master_release_versions(
 release_id,
 options = list(),
 token = NA
)
}
\arguments{
\item{master_id}{Integer value representing a valid master ID}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Get versions of the given master release
}
\details{
This function supports Pagination parameters. Below is a list of available fields.
\itemize{
\item{ \strong{page:} }{integer, the page you want to request; \strong{example}: 1}
\item{ \strong{per_page:} }{integer, the number of items per page; \strong{example}: 50}
\item{ \strong{format:} }{string, the format to filter; \strong{example}: "Vinyl"}
\item{ \strong{label:} }{string, the label to filter; \strong{example}: "Scorpio Music"}
\item{ \strong{released:} }{string, the release year to filter; \strong{example}: "1992"}
\item{ \strong{country:} }{string, the country to filter; \strong{example}: "Belgium"}
\item{ \strong{sort:} }{string, sort items by these fields: released, title, format, label, catno, country}
\item{ \strong{sort_order:} }{string, sort items in a particular order; \strong{example}: "asc"}
}
}
\examples{
# Retrieve the master versions
stats <- get_master_release_versions(1000)

# Retrieve the master versions with optional parameters
stats <- get_master_release_versions(1000,
        options = list("page" = 1, "per_page" = 50, "format" = "Vinyl"))
}
