% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_order}
\alias{get_order}
\title{Get a seller’s Marketplace order}
\usage{
get_order(
  order_id,
  token
)
}
\arguments{
\item{order_id}{The ID of the order you are fetching}

\item{token}{Token object obtained from authorize()}
}
\description{
Get a seller’s Marketplace order
}
\details{
View the data associated with an order.
Authentication as the seller is required.
}
\examples{

token <- authorize("key", "secret")
order <- get_order(1, token)
}
