% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_price_suggestions}
\alias{get_price_suggestions}
\title{Retrieve price suggestions for the provided Release ID}
\usage{
get_price_suggestions(
  release_id,
  token
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Retrieve price suggestions for the provided Release ID
}
\details{
Retrieve price suggestions for the provided Release ID. If no suggestions are available, an empty object will be returned.
Authentication is required, and the user needs to have filled out their seller settings.
Suggested prices will be denominated in the user’s selling currency.
}
\examples{
token <- authorize("key", "secret")
}
