% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identity.R
\name{get_profile}
\alias{get_profile}
\title{Retrieve a user by username.}
\usage{
get_profile(
 username,
 token
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}
}
\description{
Retrieve a user by username.
}
\details{
If authenticated as the requested user, the email key will be visible,
and the num_list count will include the user’s private lists.
If authenticated as the requested user or the user’s collection/wantlist is public,
the num_collection / num_wantlist keys will be visible.
}
\examples{
token <- authorize("key", "secret")
profile <- get_profile("name", token)
}
