% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{get_recent_uploads}
\alias{get_recent_uploads}
\title{Get recent your uploads}
\usage{
get_recent_uploads(
 token
)
}
\arguments{
\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get recent your uploads
}
\examples{
token <- authorize("key", "secret")
recent_uploads <- get_recent_uploads(token)
}
