% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_release}
\alias{get_release}
\title{Retrieve release info by a given ID}
\usage{
get_release(
 release_id,
 options = list(),
 token = NA
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{options}{Named list containing optional params.}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Retrieve release info by a given ID
}
\examples{
# Get a release
release <- get_release(1)

# Retrieve marketplace data in Euros
release <- get_release(1, options=list("curr_abbr" = "EUR"))

}
