% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_release_rating_by_user}
\alias{get_release_rating_by_user}
\title{Retrieve rating of a release by a user}
\usage{
get_release_rating_by_user(
 release_id,
 username,
 token = NA
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{username}{String containing a valid username}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Retrieve rating of a release by a user
}
\examples{
# Get a rating
rating <- get_release_rating_by_user(1, "username")
}
