% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_release_statistics}
\alias{get_release_statistics}
\title{Retrieve marketplace statistics for the provided Release ID}
\usage{
get_release_statistics(
  release_id,
  options = list(),
  token = NA
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Retrieve marketplace statistics for the provided Release ID
}
\details{
Retrieve marketplace statistics for the provided Release ID.
These statistics reflect the state of the release in the marketplace currently,
and include the number of items currently for sale, lowest listed price of any item for sale,
and whether the item is blocked for sale in the marketplace.
Authentication is optional.
Authenticated users will by default have the lowest currency expressed in their own buyer currency,
configurable in buyer settings, in the absence of the curr_abbr query parameter to specify the currency.
Unauthenticated users will have the price expressed in US Dollars, if no curr_abbr is provided.
Releases that have no items for sale in the marketplace will return a body with null data in the lowest_price and num_for_sale keys.
Releases that are blocked for sale will also have null data for these keys.
}
