% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_release_stats}
\alias{get_release_stats}
\title{Get release statistics}
\usage{
get_release_stats(
 release_id,
 token = NA
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Get release statistics
}
\examples{
# Retrieve the release statistics
stats <- get_release_stats(1)
}
