% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{get_search}
\alias{get_search}
\title{Get release statistics}
\usage{
get_search(
 token,
 query = "",
 options = list()
)
}
\arguments{
\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{query}{(optional) String value of query to search}

\item{options}{(optional) Parameters, see Details}
}
\description{
Get release statistics
}
\details{
This function supports many parameters. Below is a list of available fields, all fields are of type \strong{string}.
\itemize{
\item{ \strong{type:} }{One of \code{release}, \code{master}, \code{artist} or \code{label}; \strong{example}: "release"}
\item{ \strong{title:} }{Search by combined “Artist Name - Release Title” title field; \strong{example}: nirvana - nevermind}
\item{ \strong{release_title:} }{Search release titles; \strong{example}: "nevermind"}
\item{ \strong{credit:} }{Search release credits; \strong{example}: "kurt"}
\item{ \strong{artist:} }{Search artist names; \strong{example}: "nirvana"}
\item{ \strong{anv:} }{Search artist ANV; \strong{example}: "nirvana"}
\item{ \strong{label:} }{Search label names; \strong{example}: "dgc" }
\item{ \strong{genre:} }{Search genres; \strong{example}: "rock"}
\item{ \strong{style:} }{Search styles; \strong{example}: "grunge"}
\item{ \strong{country:} }{Search release country; \strong{example}: "canada"}
\item{ \strong{year:} }{Search release year; \strong{example}: "1991"}
\item{ \strong{format:} }{Search formats; \strong{example}: "album"}
\item{ \strong{catno:} }{Search catalog number; \strong{example}: " DGCD-24425"}
\item{ \strong{barcode:} }{Search barcodes; \strong{example}: "7 2064-24425-2 4"}
\item{ \strong{track:} }{Search track titles; \strong{example}: "smells like teen spirit"}
\item{ \strong{submitter:} }{Search submitter username; \strong{example}: "milKt"}
\item{ \strong{contributor:} }{Search contributor usernames; \strong{example}: "jerome99"}
}
}
\examples{
# Retrieve the release statistics
token <- authorize("key", "secret")
results <- get_search(token, "nirvana", options=list("year" = "1991"))

}
