% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{get_upload}
\alias{get_upload}
\title{Get details about the status of an inventory upload.}
\usage{
get_upload(
 id,
 token = NA
)
}
\arguments{
\item{id}{upload identifier}

\item{token}{Token object obtained from authorize()}
}
\description{
Get details about the status of an inventory upload.
}
\examples{
status <- get_upload(1)
}
