% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Userlist.R
\name{get_user_lists}
\alias{get_user_lists}
\title{Get a user's lists}
\usage{
get_user_lists(
 username,
 token = NA,
 options = list()
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize()}

\item{options}{(optional) List of named parameters, see Details}
}
\description{
Get a user's lists
}
\details{

}
