% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wantlist.R
\name{get_wantlist}
\alias{get_wantlist}
\title{Get a list of releases in a user's wantlist.}
\usage{
get_wantlist(
 username,
 options = list(),
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{options}{(optional) List of named parameters, see Details}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Get a list of releases in a user's wantlist.
}
\details{
Returns the list of releases in a user’s wantlist. Accepts Pagination parameters.
Basic information about each release is provided, suitable for display in a list. For detailed information, make another API call to fetch the corresponding release.
If the wantlist has been made private by its owner, you must be authenticated as the owner to view it.
The notes field will be visible if you are authenticated as the wantlist owner.
}
\examples{
wantlist <- get_wantlist("username")

wantlist <- get_want_list("username", options = list("page" = 2))
}
