% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{parse_options}
\alias{parse_options}
\title{parse optional parameters}
\usage{
parse_options(
 options
)
}
\arguments{
\item{options}{Named list containing a key - value combination}
}
\description{
parse optional parameters
}
\details{
This function is not supposed to used directly, but through other functions instead
}
