% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{post}
\alias{post}
\title{Post request to the discogs server}
\usage{
post(
 url,
 content,
 token = NA
)
}
\arguments{
\item{url}{full URL to API endpoint}

\item{content}{The content to post / upload}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}
}
\description{
Post request to the discogs server
}
\details{
This function is not supposed to used directly, but through other functions instead
}
