% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{put}
\alias{put}
\title{Put request to the discogs server}
\usage{
put(
 url,
 content,
 token
)
}
\arguments{
\item{token}{Token object obtained from authorize()}
}
\description{
Put request to the discogs server
}
\details{

}
\examples{
token <- authorize("key", "secret")
}
