% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{put_release_rating}
\alias{put_release_rating}
\title{Submit a rating for a release}
\usage{
put_release_rating(
 release_id,
 username,
 rating
 token
)
}
\arguments{
\item{release_id}{Integer value representing a valid release ID}

\item{username}{String containing a valid username}

\item{rating}{Integer value between 0-5}

\item{token}{Token object obtained from authorize()}
}
\description{
Submit a rating for a release
}
\examples{
# Submit a rating
token <- authorize("key", "secret")
put_release_rating(1, "username", 5, token)
}
