% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wantlist.R
\name{add_to_wantlist}
\alias{add_to_wantlist}
\title{Add a release to your wantlist}
\usage{
add_to_wantlist(
 username,
 release_id,
 token,
 options = list()
)
}
\arguments{
\item{username}{String containing a valid username}

\item{release_id}{Integer value representing a valid release ID}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{options}{(optional) List of named parameters, see Details}
}
\description{
Add a release to your wantlist
}
\details{
Add a release to a user’s wantlist.
Authentication as the wantlist owner is required.
#'## Parameters
\itemize{
\item \strong{notes}: User notes to associate with this release.
\item \strong{rating}: User’s rating of this release, from 0 (unrated) to 5 (best). Defaults to 0.
}
}
\examples{
token <- authorize("key", "secret")
add_to_wantlist("username", 1000, token)

add_to_wantlist("username", 1000, token, options = list(notes = "B-side is really cool", rating = 4))
}
