% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{create_collection_folder}
\alias{create_collection_folder}
\title{Create a new folder in your collection}
\usage{
create_collection_folder(
 username,
 token,
 name = ""
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{name}{(optional) Name for the folder}
}
\description{
Create a new folder in your collection
}
\details{
Authentication as the collection owner is required.
}
\examples{
token <- authorize("key", "secret")
create_collection_folder("username", token, "my_folder")
}
