% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{delete}
\alias{delete}
\title{Delete request to the discogs server}
\usage{
delete(
 url,
 token
)
}
\arguments{
\item{token}{Token object obtained from authorize()}
}
\description{
Delete request to the discogs server
}
\details{

}
\examples{
token <- authorize("key", "secret")
}
