% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory.R
\name{download_export}
\alias{download_export}
\title{Download an export}
\usage{
download_export(
 export_id,
 token
)
}
\arguments{
\item{export_id}{Id of the export}

\item{token}{Token object obtained from authorize()}
}
\description{
Download an export
}
\details{

}
\examples{
token <- authorize("key", "secret")
csv <- download_export(1, token)
}
