% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{edit_field_instance}
\alias{edit_field_instance}
\title{Edit fields of an instance}
\usage{
edit_field_instance(
 username
 value
 folder_id
 release_id
 instance_id
 field_id
 token
)
}
\arguments{
\item{username}{String containing a valid username}

\item{value}{The new value of the field}

\item{folder_id}{The ID of the folder to modify}

\item{release_id}{Integer value representing a valid release ID}

\item{instance_id}{The ID of the instance}

\item{field_id}{The ID of the field}

\item{token}{Token object obtained from authorize()}
}
\description{
Edit fields of an instance
}
\details{

}
