% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marketplace.R
\name{get_list_orders}
\alias{get_list_orders}
\title{Returns a list of the authenticated user’s orders}
\usage{
get_list_orders(
  token,
  options = list(),
)
}
\arguments{
\item{token}{Token object obtained from authorize() or a string containing your personal access token}

\item{options}{(optional) Parameters, see Details}
}
\description{
Returns a list of the authenticated user’s orders
}
\details{
Returns a list of the authenticated user’s orders. Accepts Pagination parameters.
}
\examples{
token <- authorize("key", "secret")
}
