% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identity.R
\name{get_user_contributions}
\alias{get_user_contributions}
\title{Retrieve a user’s contributions by username.}
\usage{
get_user_contributions(
 username,
 options = list(),
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{options}{(optional) Parameters, see Details}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Retrieve a user’s contributions by username.
}
\details{
Accepts Pagination parameters.
}
