% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Userlist.R
\name{get_user_lists}
\alias{get_user_lists}
\title{Get a user's lists}
\usage{
get_user_lists(
 username,
 token = NA,
 options = list()
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{(optional) Token object obtained from authorize() or a string containing your personal access token}

\item{options}{(optional) List of named parameters, see Details}
}
\description{
Get a user's lists
}
\details{
Returns a User’s Lists. Private Lists will only display when authenticated as the owner. Accepts Pagination parameters.
\subsection{Pagination options;}{
\itemize{
\item \strong{page:} integer, the page you want to request; \strong{example}: 1
\item \strong{per_page:} integer, the number of items per page; \strong{example}: 50
}
}
}
\examples{
user_list <- get_user_lists("username")
}
