% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identity.R
\name{get_user_submissions}
\alias{get_user_submissions}
\title{Retrieve a user’s submissions by username.}
\usage{
get_user_submissions(
 username,
 token = NA
)
}
\arguments{
\item{username}{String containing a valid username}

\item{token}{(optional) Token object obtained from authorize()}
}
\description{
Retrieve a user’s submissions by username.
}
\details{
Accepts Pagination parameters.
}
