% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{post}
\alias{post}
\title{Post request to the discogs server}
\usage{
post(
 url,
 content,
 token = NA
)
}
\arguments{
\item{token}{Token object obtained from authorize()}
}
\description{
Post request to the discogs server
}
\details{

}
\examples{
token <- authorize("key", "secret")
}
